refactor: created DataUploadButton component
